t2726: remove 5 redundant session-rename tool schema leaks
session-rename-guards.ts was in .opencode/tool/ causing OpenCode to expose
isDefaultBranchTitle, isTitleOverwritable, and syncSessionWithBranch as
model-callable tools. These are internal helpers, not user-facing operations.

Fixes:
- Move session-rename-guards.ts to .opencode/lib/ (not scanned for tools)
- Update import path in session-rename.ts: ./session-rename-guards -> ../lib/
- Remove re-exports of isDefaultBranchTitle/isTitleOverwritable from session-rename.ts
- Remove export from syncSessionWithBranch (called internally by sync_branch only)
- Update test import path to match new lib/ location

Result: 7 session-rename tools -> 2 (session-rename + session-rename_sync_branch)
