+++ .opencode/lib/session-rename-guards.ts
